Rewrite how BackupJob state reconciliation works
Intermediate stopping point in refactor. Attempts to move all domain
logic into an InternalState struct which holds all the relevant state
for K8s necessary to reconcile the CRD. Move the fetching of this state
to its own function, and also move the derivation of conditions & phase
to their own functions.
This tries to divide the fetching of state from the cluster from the
logic around how to reconcile the resource given that state.

The next step will be to simplify as much as possible, and remove
duplicate domain logic expressed in the derivation of conditions and
reconciliation of the resource.