Add prune/check support for BackupSchedule
Pull out the creation of a Job resource from BackupJob into a separate
module, which can then be used by both BackupJob & BackupSchedule to
create each of the three job types: Prune, Check, and Backup. The first
two are done by BackupSchedule, the latter by BackupJob. Move
ensure_rbac over too, as it's only used when creating a job.

Refactor BackupSchedule to use conditions to track state of each of the
three jobs types; use conditions to derive a state to show to the user.

Add support for checking and purging to Walle worker.